code/
Symbol detection can be made easier if we know the best number of Gaussians for each symbol. Instead of
using the same number of Gaussians for all symbols, the system can be fine-tuned for each symbol, which
could make symbol detection more accurate and efficient. There is a clear ”elbow” point in many of the plots. This is a key point where the rate of decrease in
BIC starts to slow down. Up until now, adding more Gaussians has made a big difference in how well the
model fits the data. Once we get past the elbow, the extra complexity (more Gaussians) doesn’t improve
the fit in a proportional way. The location of the elbow point is different for each symbol, which shows how
different and complicated the data that goes with each symbol is. If a symbol needs more Gaussians to get
to the elbow, it might be because it is more complicated or noisy than others. Once we have these BIC plots, we can fit GMMs with the best number of Gaussians for each symbol. A Gaussian Mixture Model (GMM) has been estimated for every unique symbol present in the data set.
The list of (.parquet) files in the designated directory or location of the provided data set has been declared. Each unique symbol is shown by a different
Gaussian mixture. This shows the assumed chance distributions for the symbols that were received. There
is a clear overlap between these Gaussian mixtures. This shows that there is inherent ambiguity and the
possibility of symbol confusion because the probability distributions cross in some places. This overlap
shows how hard it is to correctly classify and understand the symbols that are received. This is why reliable
communication requires advanced decoding and error-correction strategies.
For each unique symbol (or label) in the point label column of the data set, the data for that symbol
is taken out. This data is used to train a set of Gaussian Mixture Models with anywhere from 1 to 5
components. For each of these models, the Bayesian Information Criterion (BIC) is worked out. The GMM
with the lowest BIC is thought to be the best for that sign. This method tries to find the model that fits
the data best with the fewest assumptions. For each data point in the collection, the likelihood of that data point belonging to each symbol’s GMM
is calculated. The predicted label for that data point is the symbol that matches the GMM with the
best likelihood. A new column called predicted label holds the predicted labels. Finally, the true labels
(point label) and the predicted labels (predicted label) are printed for each data point. The probabilities of each data point are being computed, and the predictions are being made in accordance
with those results. This gives a deeper insight into reasoning why the data point should be associated with
a certain symbol.
The ratio of right predictions to all predictions is used to figure out how accurate a model is. The green data points provide valuable information regarding the model’s accurate predictions. The
distribution of these points is dense in the vicinity of the corresponding 16-QAM constellation points, which
indicates that the model exhibits a generally accurate performance.